@jeffbolznv 2025-02-02
> See branch https://github.com/remyoudompheng/llama.cpp/tree/vulkan-iq-mmv for MMV kernels for IQ2 and IQ3 quants

Very cool! I tested this and it's functionally correct and perf is better on RTX 4070.

I dug into the perf a bit and realized that a significant amount of time is spent in init_iq_shmem since the LUT is so large. I think I had suggested this before, but this more unrollable loop code helps:

```
    [[unroll]] for (uint i = 0; i < iq2s_grid.length(); i += wgsize.x) {
        iq2s_grid[i + gl_LocalInvocationIndex.x] = iq2s_grid_const[i + gl_LocalInvocationIndex.x];
    }
```

Even then, it's still expensive and that suggests we should be doing more work per workgroup to amortize the cost. The large shared memory allocation may also limit the number of workgroups that can run concurrently, which argues for using larger workgroups. I verified that increasing NUM_ROWS and workgroup size helps:

```
before
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 11076 runs -    95.35 us/run - 117.44 MFLOP/run -   1.23 TFLOPS
tot iq-mmv branch
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 12780 runs -    80.79 us/run - 117.44 MFLOP/run -   1.45 TFLOPS
unroll iq_shmem
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 14484 runs -    73.06 us/run - 117.44 MFLOP/run -   1.61 TFLOPS
rm_kq=4
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 17040 runs -    61.34 us/run - 117.44 MFLOP/run -   1.91 TFLOPS
rm_kq=4 and wgsize=64
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 17892 runs -    56.95 us/run - 117.44 MFLOP/run -   2.06 TFLOPS
rm_kq=8 and wgsize=64
  MUL_MAT(type_a=iq2_s,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 19596 runs -    51.98 us/run - 117.44 MFLOP/run -   2.26 TFLOPS
```

You don't need to do all of this at once. I think the unrollable loops is a simple change and should help everywhere. For figuring out the best values for all the knobs we'll need to get more exhaustive data from different HW and also test with real models.